+++ pkgs/jnigen/test/simple_package_test/generate.dart
@@ -46,6 +46,8 @@ final javaFiles = [
   join(javaPrefix, 'inheritance', 'GenericDerivedClass.java'),
   join(javaPrefix, 'inheritance', 'SpecificDerivedClass.java'),
   join(javaPrefix, 'interfaces', 'GenericInterface.java'),
+  join(javaPrefix, 'interfaces', 'InheritedFromMyInterface.java'),
+  join(javaPrefix, 'interfaces', 'InheritedFromMyRunnable.java'),
   join(javaPrefix, 'interfaces', 'MyInterface.java'),
   join(javaPrefix, 'interfaces', 'MyInterfaceConsumer.java'),
   join(javaPrefix, 'interfaces', 'MyRunnable.java'),