+++ pkgs/jnigen/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.14.1-wip
+
 - Added support for generating matching Kotlin operators as Dart operators.
+- Include the methods of the superinterfaces of a class or interface in the
+  bindings.
 
 ## 0.14.0